Add referenced routes and models in web.php
Code is added in web.php that references various table routes for search, CRUD actions, and alerts. Additional lines of code indicate the creation of various new models for the 'veicolo' object. These changes pave the way for future implementation of more complex interactions with the 'veicolo' object in the web application.