Replace Laravel Sanctum with Passport for API Authentication
Historically, Laravel Sanctum was used for API Authentication in the project. However, the application need changes, so this commit replaces Laravel Sanctum with Laravel Passport. Laravel Passport provides a full Oauth2 server implementation which is required for the application. The key changes implemented also include the update of some existing files and deletion of Sanctum related files.