Remove BaseModel class from app models
The BaseModel class has been removed from the 'app/Models' directory. This class provided common functions and rules for validation for other models and had a significant amount of code. It's possible that this class was deleted as a part of a refactoring process to simplify or optimize the application structure.